+++ pode.build.ps1
@@ -4,6 +4,17 @@ param (
 )
 
 <#
+# Dependency Versions
+#>
+
+$PesterVersion = '4.8.0'
+$MkDocsVersion = '1.0.4'
+$CoverallsVersion = '1.0.25'
+$7ZipVersion = '18.5.0.20180730'
+$ChecksumVersion = '0.2.0'
+$MkDocsThemeVersion = '4.2.0'
+
+<#
 # Helper Functions
 #>
 
@@ -92,28 +103,28 @@ task ChocoDeps -If (Test-IsWindows) {
 # Synopsis: Install dependencies for packaging
 task PackDeps -If (Test-IsWindows) ChocoDeps, {
     if (!(Test-Command 'checksum')) {
-        Invoke-Install 'checksum' '0.2.0'
+        Invoke-Install 'checksum' $ChecksumVersion
     }
 
     if (!(Test-Command '7z')) {
-        Invoke-Install '7zip' '18.5.0.20180730'
+        Invoke-Install '7zip' $7ZipVersion
     }
 }
 
 # Synopsis: Install dependencies for running tests
 task TestDeps {
     # install pester
-    if (((Get-Module -ListAvailable Pester) | Where-Object { $_.Version -ieq '4.4.2' }) -eq $null) {
+    if (((Get-Module -ListAvailable Pester) | Where-Object { $_.Version -ieq $PesterVersion }) -eq $null) {
         Write-Host 'Installing Pester'
-        Install-Module -Name Pester -Scope CurrentUser -RequiredVersion '4.4.2' -Force -SkipPublisherCheck
+        Install-Module -Name Pester -Scope CurrentUser -RequiredVersion $PesterVersion -Force -SkipPublisherCheck
     }
 
     # install coveralls
     if (Test-IsAppVeyor)
     {
-        if (((Get-Module -ListAvailable coveralls) | Where-Object { $_.Version -ieq '1.0.25' }) -eq $null) {
+        if (((Get-Module -ListAvailable coveralls) | Where-Object { $_.Version -ieq $CoverallsVersion }) -eq $null) {
             Write-Host 'Installing Coveralls'
-            Install-Module -Name coveralls -Scope CurrentUser -RequiredVersion '1.0.25' -Force -SkipPublisherCheck
+            Install-Module -Name coveralls -Scope CurrentUser -RequiredVersion $CoverallsVersion -Force -SkipPublisherCheck
         }
     }
 }
@@ -121,11 +132,12 @@ task TestDeps {
 # Synopsis: Install dependencies for documentation
 task DocsDeps ChocoDeps, {
     if (!(Test-Command 'mkdocs')) {
-        Invoke-Install 'mkdocs' '1.0.4'
+        Invoke-Install 'mkdocs' $MkDocsVersion
     }
 
-    if ((pip list --format json --disable-pip-version-check | ConvertFrom-Json).name -inotcontains 'mkdocs-material') {
-        pip install mkdocs-material --disable-pip-version-check
+    $_installed = (pip list --format json --disable-pip-version-check | ConvertFrom-Json)
+    if (($_installed | Where-Object { $_.name -ieq 'mkdocs-material' -and $_.version -ieq $MkDocsThemeVersion } | Measure-Object).Count -eq 0) {
+        pip install "mkdocs-material==$($MkDocsThemeVersion)" --force-reinstall --disable-pip-version-check
     }
 }
 
@@ -155,8 +167,8 @@ task Pack -If (Test-IsWindows) 7Zip, ChocoPack
 # Synopsis: Run the tests
 task Test TestDeps, {
     $p = (Get-Command Invoke-Pester)
-    if ($null -eq $p -or $p.Version -ine '4.4.2') {
-        Import-Module Pester -Force -RequiredVersion '4.4.2'
+    if ($null -eq $p -or $p.Version -ine $PesterVersion) {
+        Import-Module Pester -Force -RequiredVersion $PesterVersion
     }
 
     $Script:TestResultFile = "$($pwd)/TestResults.xml"