+++ src/Tools/Helpers.ps1
@@ -1028,8 +1028,8 @@ function Join-PodePaths
 
     # if there are any more, add them on
     if ($Paths.Length -gt 2) {
-        $Paths[2..($Paths.Length - 1)] | ForEach-Object {
-            $_path = Join-Path $_path $_
+        foreach ($p in $Paths[2..($Paths.Length - 1)]) {
+            $_path = Join-Path $_path $p
         }
     }
 
@@ -1210,11 +1210,15 @@ function ConvertFrom-PodeRequestContent
 {
     param (
         [Parameter()]
-        $Request
+        $Request,
+
+        [Parameter()]
+        [string]
+        $ContentType
     )
 
     # get the requests content type and boundary
-    $MetaData = Get-PodeContentTypeAndBoundary -ContentType $Request.ContentType
+    $MetaData = Get-PodeContentTypeAndBoundary -ContentType $ContentType
     $Encoding = $Request.ContentEncoding
 
     # result object for data/files
@@ -1241,7 +1245,12 @@ function ConvertFrom-PodeRequestContent
     # run action for the content type
     switch ($MetaData.ContentType) {
         { $_ -ilike '*/json' } {
-            $Result.Data = ($Content | ConvertFrom-Json)
+            if (Test-IsPSCore) {
+                $Result.Data = ($Content | ConvertFrom-Json -AsHashtable)
+            }
+            else {
+                $Result.Data = ($Content | ConvertFrom-Json)
+            }
         }
 
         { $_ -ilike '*/xml' } {
@@ -1366,8 +1375,8 @@ function ConvertFrom-PodeNameValueToHashTable
     }
 
     $ht = @{}
-    $Collection.Keys | ForEach-Object {
-        $ht[$_] = $Collection[$_]
+    foreach ($key in $Collection.Keys) {
+        $ht[$key] = $Collection[$key]
     }
 
     return $ht
@@ -1499,12 +1508,31 @@ function Convert-PodePathPatternToRegex
     param (
         [Parameter()]
         [string]
-        $Path
+        $Path,
+
+        [switch]
+        $NotSlashes,
+
+        [switch]
+        $NotStrict
     )
 
     $Path = $Path -ireplace '\.', '\.'
-    $Path = $Path -ireplace '[\\/]', '[\\/]'
+
+    if (!$NotSlashes) {
+        if ($Path -match '[\\/]\*$') {
+            $Path = $Path -replace '[\\/]\*$', '/{0,1}*'
+        }
+
+        $Path = $Path -ireplace '[\\/]', '[\\/]'
+    }
+
     $Path = $Path -ireplace '\*', '.*?'
+
+    if ($NotStrict) {
+        return $Path
+    }
+
     return "^$($Path)$"
 }
 
@@ -1516,7 +1544,10 @@ function Convert-PodePathPatternsToRegex
         $Paths,
 
         [switch]
-        $NotSlashes
+        $NotSlashes,
+
+        [switch]
+        $NotStrict
     )
 
     # remove any empty entries
@@ -1532,22 +1563,18 @@ function Convert-PodePathPatternsToRegex
     # replace certain chars
     $Paths = @($Paths | ForEach-Object {
         if (!(Test-Empty $_)) {
-            $tmp = $_ -ireplace '\.', '\.'
-
-            if (!$NotSlashes) {
-                if ($tmp -match '[\\/]\*$') {
-                    $tmp = $tmp -replace '[\\/]\*$', '/{0,1}*'
-                }
-
-                $tmp = $tmp -ireplace '[\\/]', '[\\/]'
-            }
-
-            $tmp -ireplace '\*', '.*?'
+            Convert-PodePathPatternToRegex -Path $_ -NotStrict -NotSlashes:$NotSlashes
         }
     })
 
     # join them all together
-    return "^($($Paths -join '|'))$"
+    $joined = "($($Paths -join '|'))"
+
+    if ($NotStrict) {
+        return "$($joined)"
+    }
+
+    return "^$($joined)$"
 }
 
 function Get-PodeModulePath
@@ -1581,4 +1608,286 @@ function Get-PodeModulePath
 function Get-PodeModuleRootPath
 {
     return (Split-Path -Parent -Path $PodeContext.Server.PodeModulePath)
+}
+
+function Get-PodeUrl
+{
+    return "$($WebEvent.Protocol)://$($WebEvent.Endpoint)$($WebEvent.Path)"
+}
+
+function Find-PodeErrorPage
+{
+    param (
+        [Parameter()]
+        [int]
+        $Code,
+
+        [Parameter()]
+        [string]
+        $ContentType
+    )
+
+    # if a defined content type is supplied, attempt to find an error page for that first
+    if (!(Test-Empty $ContentType)) {
+        $path = Get-PodeErrorPage -Code $Code -ContentType $ContentType
+        if (!(Test-Empty $path)) {
+            return @{ 'Path' = $path; 'ContentType' = $ContentType }
+        }
+    }
+
+    # if a defined route error page content type is supplied, attempt to find an error page for that
+    if (!(Test-Empty $WebEvent.ErrorType)) {
+        $path = Get-PodeErrorPage -Code $Code -ContentType $WebEvent.ErrorType
+        if (!(Test-Empty $path)) {
+            return @{ 'Path' = $path; 'ContentType' = $WebEvent.ErrorType }
+        }
+    }
+
+    # if route patterns have been defined, see if an error content type matches and attempt that
+    if (!(Test-Empty $PodeContext.Server.Web.ErrorPages.Routes)) {
+        # find type by pattern
+        $matched = ($PodeContext.Server.Web.ErrorPages.Routes.Keys | Where-Object {
+            $WebEvent.Path -imatch $_
+        } | Select-Object -First 1)
+
+        # if we have a match, see if a page exists
+        if (!(Test-Empty $matched)) {
+            $type = $PodeContext.Server.Web.ErrorPages.Routes[$matched]
+            $path = Get-PodeErrorPage -Code $Code -ContentType $type
+            if (!(Test-Empty $path)) {
+                return @{ 'Path' = $path; 'ContentType' = $type }
+            }
+        }
+    }
+
+    # if we're using strict typing, attempt that, if we have a content type
+    if ($PodeContext.Server.Web.ErrorPages.StrictContentTyping -and !(Test-Empty $WebEvent.ContentType)) {
+        $path = Get-PodeErrorPage -Code $Code -ContentType $WebEvent.ContentType
+        if (!(Test-Empty $path)) {
+            return @{ 'Path' = $path; 'ContentType' = $WebEvent.ContentType }
+        }
+    }
+
+    # if we have a default defined, attempt that
+    if (!(Test-Empty $PodeContext.Server.Web.ErrorPages.Default)) {
+        $path = Get-PodeErrorPage -Code $Code -ContentType $PodeContext.Server.Web.ErrorPages.Default
+        if (!(Test-Empty $path)) {
+            return @{ 'Path' = $path; 'ContentType' = $PodeContext.Server.Web.ErrorPages.Default }
+        }
+    }
+
+    # if there's still no error page, use default HTML logic
+    $type = Get-PodeContentType -Extension 'html'
+    $path = (Get-PodeErrorPage -Code $Code -ContentType $type)
+
+    if (!(Test-Empty $path)) {
+        return @{ 'Path' = $path; 'ContentType' = $type }
+    }
+
+    return $null
+}
+
+function Get-PodeErrorPage
+{
+    param (
+        [Parameter()]
+        [int]
+        $Code,
+
+        [Parameter()]
+        [string]
+        $ContentType
+    )
+
+    # parse the passed content type
+    $ContentType = (Get-PodeContentTypeAndBoundary -ContentType $ContentType).ContentType
+
+    # object for the page path
+    $path = $null
+
+    # attempt to find a custom error page
+    $path = Find-PodeCustomErrorPage -Code $Code -ContentType $ContentType
+
+    # if there's no custom page found, attempt to find an inbuilt page
+    if (Test-Empty $path) {
+        $podeRoot = Join-Path (Get-PodeModuleRootPath) 'Misc'
+        $path = Find-PodeFileForContentType -Path $podeRoot -Name 'default-error-page' -ContentType $ContentType -Engine 'pode'
+    }
+
+    # if there's no path found, or it's inaccessible, return null
+    if (!(Test-PodePath $path -NoStatus)) {
+        return $null
+    }
+
+    return $path
+}
+
+function Find-PodeCustomErrorPage
+{
+    param (
+        [Parameter()]
+        [int]
+        $Code,
+
+        [Parameter()]
+        [string]
+        $ContentType
+    )
+
+    # get the custom errors path
+    $customErrPath = $PodeContext.Server.InbuiltDrives['errors']
+
+    # if there's no custom error path, return
+    if (Test-Empty $customErrPath) {
+        return $null
+    }
+
+    # retrieve a status code page
+    $path = (Find-PodeFileForContentType -Path $customErrPath -Name "$($Code)" -ContentType $ContentType)
+    if (!(Test-Empty $path)) {
+        return $path
+    }
+
+    # retrieve default page
+    $path = (Find-PodeFileForContentType -Path $customErrPath -Name 'default' -ContentType $ContentType)
+    if (!(Test-Empty $path)) {
+        return $path
+    }
+
+    # no file was found
+    return $null
+}
+
+function Find-PodeFileForContentType
+{
+    param (
+        [Parameter()]
+        [string]
+        $Path,
+
+        [Parameter()]
+        [string]
+        $Name,
+
+        [Parameter()]
+        [string]
+        $ContentType,
+
+        [Parameter()]
+        [string]
+        $Engine = $null
+    )
+
+    # get all files at the path that start with the name
+    $files = @(Get-ChildItem -Path (Join-Path $Path "$($Name).*"))
+
+    # if there are no files, return
+    if (Test-Empty $files) {
+        return $null
+    }
+
+    # filter the files by the view engine extension (but only if the current engine is dynamic - non-html)
+    if ((Test-Empty $Engine) -and $PodeContext.Server.ViewEngine.IsDynamic) {
+        $Engine = $PodeContext.Server.ViewEngine.Extension
+    }
+
+    $Engine = (coalesce $Engine 'pode')
+    if ($Engine -ine 'pode') {
+        $Engine = "($($Engine)|pode)"
+    }
+
+    $engineFiles = @($files | Where-Object { $_.Name -imatch "\.$($Engine)$" })
+    $files = @($files | Where-Object { $_.Name -inotmatch "\.$($Engine)$" })
+
+    # only attempt static files if we still have files after any engine filtering
+    if (!(Test-Empty $files))
+    {
+        # get files of the format '<name>.<type>'
+        $file = ($files | Where-Object {
+            if ($_.Name -imatch "^$($Name)\.(?<ext>.*?)$") {
+                return ($ContentType -ieq (Get-PodeContentType -Extension $Matches['ext']))
+            }
+        } | Select-Object -First 1)
+
+        if (!(Test-Empty $file)) {
+            return $file.FullName
+        }
+    }
+
+    # only attempt these formats if we have a files for the view engine
+    if (!(Test-Empty $engineFiles))
+    {
+        # get files of the format '<name>.<type>.<engine>'
+        $file = ($engineFiles | Where-Object {
+            if ($_.Name -imatch "^$($Name)\.(?<ext>.*?)\.$($engine)$") {
+                return ($ContentType -ieq (Get-PodeContentType -Extension $Matches['ext']))
+            }
+        } | Select-Object -First 1)
+
+        if (!(Test-Empty $file)) {
+            return $file.FullName
+        }
+
+        # get files of the format '<name>.<engine>'
+        $file = ($engineFiles | Where-Object {
+            return ($_.Name -imatch "^$($Name)\.$($engine)$")
+        } | Select-Object -First 1)
+
+        if (!(Test-Empty $file)) {
+            return $file.FullName
+        }
+    }
+
+    # no file was found
+    return $null
+}
+
+function Test-PodeRelativePath
+{
+    param (
+        [Parameter(Mandatory=$true)]
+        [string]
+        $Path
+    )
+
+    if (@('.', '..') -contains $Path) {
+        return $true
+    }
+
+    return ($Path -match '^\.{1,2}[\\/]')
+}
+
+function Get-PodeRelativePath
+{
+    param (
+        [Parameter(Mandatory=$true)]
+        [string]
+        $Path,
+
+        [Parameter()]
+        [string]
+        $RootPath,
+
+        [switch]
+        $JoinRoot,
+
+        [switch]
+        $Resolve
+    )
+
+    # if the path is relative, join to root if flagged
+    if ($JoinRoot -and (Test-PodeRelativePath -Path $Path)) {
+        if (Test-Empty $RootPath) {
+            $RootPath = $PodeContext.Server.Root
+        }
+
+        $Path = Join-Path $RootPath $Path
+    }
+
+    # if flagged, resolve the path
+    if ($Resolve) {
+        $Path = (Resolve-Path -Path $Path -ErrorAction Ignore).Path
+    }
+
+    return $Path
 }