+++ tests/unit/Tools/Helpers.Tests.ps1
@@ -615,10 +615,9 @@ Describe 'ConvertFrom-PodeRequestContent' {
             $value = '<root><value>test</value></root>'
             Mock Read-PodeStreamToEnd { return $value }
 
-            $result = ConvertFrom-PodeRequestContent @{
-                'ContentType' = 'text/xml';
+            $result = ConvertFrom-PodeRequestContent -Request @{
                 'ContentEncoding' = [System.Text.Encoding]::UTF8;
-            }
+            } -ContentType 'text/xml'
 
             $result.Data | Should Not Be $null
             $result.Data.root | Should Not Be $null
@@ -629,10 +628,9 @@ Describe 'ConvertFrom-PodeRequestContent' {
             $value = '{ "value": "test" }'
             Mock Read-PodeStreamToEnd { return $value }
 
-            $result = ConvertFrom-PodeRequestContent @{
-                'ContentType' = 'application/json';
+            $result = ConvertFrom-PodeRequestContent -Request @{
                 'ContentEncoding' = [System.Text.Encoding]::UTF8;
-            }
+            } -ContentType 'application/json'
 
             $result.Data | Should Not Be $null
             $result.Data.value | Should Be 'test'
@@ -642,10 +640,9 @@ Describe 'ConvertFrom-PodeRequestContent' {
             $value = "value`ntest"
             Mock Read-PodeStreamToEnd { return $value }
 
-            $result = ConvertFrom-PodeRequestContent @{
-                'ContentType' = 'text/csv';
+            $result = ConvertFrom-PodeRequestContent -Request @{
                 'ContentEncoding' = [System.Text.Encoding]::UTF8;
-            }
+            } -ContentType 'text/csv'
 
             $result | Should Not Be $null
             $result.Data[0].value | Should Be 'test'
@@ -655,10 +652,9 @@ Describe 'ConvertFrom-PodeRequestContent' {
             $value = "test"
             Mock Read-PodeStreamToEnd { return $value }
             
-            (ConvertFrom-PodeRequestContent @{
-                'ContentType' = 'text/custom';
+            (ConvertFrom-PodeRequestContent -Request @{
                 'ContentEncoding' = [System.Text.Encoding]::UTF8;
-            }).Data | Should Be 'test'
+            } -ContentType 'text/custom').Data | Should Be 'test'
         }
     }
 }
@@ -849,5 +845,148 @@ Describe 'Set-PodeCertificate' {
         Set-PodeCertificate -Address 'localhost' -Port 8080 -Certificate 'name' | Out-Null
 
         Assert-MockCalled Write-Host -Times 1 -Scope It
+    }
+}
+
+Describe 'Get-PodeUrl' {
+    It 'Returns a url from the web event' {
+        $WebEvent = @{
+            'Protocol' = 'http';
+            'Endpoint' = 'foo.com/';
+            'Path' = 'about'
+        }
+
+        Get-PodeUrl | Should Be 'http://foo.com/about'
+    }
+}
+
+Describe 'Convert-PodePathPatternToRegex' {
+    It 'Convert a path to regex' {
+        Convert-PodePathPatternToRegex -Path '/api*' | Should Be '^[\\/]api.*?$'
+    }
+
+    It 'Convert a path to regex non-strict' {
+        Convert-PodePathPatternToRegex -Path '/api*' -NotStrict | Should Be '[\\/]api.*?'
+    }
+
+    It 'Convert a path to regex, but not slashes' {
+        Convert-PodePathPatternToRegex -Path '/api*' -NotSlashes | Should Be '^/api.*?$'
+    }
+
+    It 'Convert a path to regex, but not slashes and non-strict' {
+        Convert-PodePathPatternToRegex -Path '/api*' -NotSlashes -NotStrict | Should Be '/api.*?'
+    }
+}
+
+Describe 'Convert-PodePathPatternsToRegex' {
+    It 'Convert paths to regex' {
+        Convert-PodePathPatternsToRegex -Paths @('/api*', '/users*') | Should Be '^([\\/]api.*?|[\\/]users.*?)$'
+    }
+
+    It 'Convert paths to regex non-strict' {
+        Convert-PodePathPatternsToRegex -Paths @('/api*', '/users*') -NotStrict | Should Be '([\\/]api.*?|[\\/]users.*?)'
+    }
+
+    It 'Convert paths to regex, but not slashes' {
+        Convert-PodePathPatternsToRegex -Paths @('/api*', '/users*') -NotSlashes | Should Be '^(/api.*?|/users.*?)$'
+    }
+
+    It 'Convert paths to regex, but not slashes and non-strict' {
+        Convert-PodePathPatternsToRegex -Paths @('/api*', '/users*') -NotSlashes -NotStrict | Should Be '(/api.*?|/users.*?)'
+    }
+}
+
+Describe 'ConvertFrom-PodeFile' {
+    It 'Generates dynamic content' {
+        $content = 'Value = $(1+1)'
+        ConvertFrom-PodeFile -Content $content | Should Be 'Value = 2'
+    }
+
+    It 'Generates dynamic content, using parameters' {
+        $content = 'Value = $($data["number"])'
+        ConvertFrom-PodeFile -Content $content -Data @{ 'number' = 3 } | Should Be 'Value = 3'
+    }
+}
+
+Describe 'Test-PodeRelativePath' {
+    It 'Returns true for .' {
+        Test-PodeRelativePath -Path '.' | Should Be $true
+    }
+
+    It 'Returns true for ..' {
+        Test-PodeRelativePath -Path '..' | Should Be $true
+    }
+
+    It 'Returns true for relative file' {
+        Test-PodeRelativePath -Path './file.txt' | Should Be $true
+    }
+
+    It 'Returns true for relative folder' {
+        Test-PodeRelativePath -Path '../folder' | Should Be $true
+    }
+
+    It 'Returns false for literal windows path' {
+        Test-PodeRelativePath -Path 'c:/path' | Should Be $false
+    }
+
+    It 'Returns false for literal nix path' {
+        Test-PodeRelativePath -Path '/path' | Should Be $false
+    }
+}
+
+Describe 'Get-PodeRelativePath' {
+    $PodeContext = @{ 'Server' = @{ 'Root' = 'c:/' } }
+
+    It 'Returns back a literal path' {
+        Mock Test-PodeRelativePath { return $false }
+        Get-PodeRelativePath -Path 'c:/path' | Should Be 'c:/path'
+    }
+
+    It 'Returns null for non-existent literal path when resolving' {
+        Mock Test-PodeRelativePath { return $false }
+        Mock Resolve-Path { return $null }
+        Get-PodeRelativePath -Path 'c:/path' -Resolve | Should Be ([string]::Empty)
+    }
+
+    It 'Returns path for literal path when resolving' {
+        Mock Test-PodeRelativePath { return $false }
+        Mock Resolve-Path { return @{ 'Path' = 'c:/path' } }
+        Get-PodeRelativePath -Path 'c:/path' -Resolve | Should Be 'c:/path'
+    }
+
+    It 'Returns back a relative path' {
+        Mock Test-PodeRelativePath { return $true }
+        Get-PodeRelativePath -Path './path' | Should Be './path'
+    }
+
+    It 'Returns null for a non-existent relative path when resolving' {
+        Mock Test-PodeRelativePath { return $true }
+        Mock Resolve-Path { return $null }
+        Get-PodeRelativePath -Path './path' -Resolve | Should Be ([string]::Empty)
+    }
+
+    It 'Returns path for a relative path when resolving' {
+        Mock Test-PodeRelativePath { return $true }
+        Mock Resolve-Path { return @{ 'Path' = 'c:/path' } }
+        Get-PodeRelativePath -Path './path' -Resolve | Should Be 'c:/path'
+    }
+
+    It 'Returns path for a relative path joined to default root' {
+        Mock Test-PodeRelativePath { return $true }
+        Mock Join-Path { return 'c:/path' }
+        Get-PodeRelativePath -Path './path' -JoinRoot | Should Be 'c:/path'
+    }
+
+    It 'Returns resolved path for a relative path joined to default root when resolving' {
+        Mock Test-PodeRelativePath { return $true }
+        Mock Join-Path { return 'c:/path' }
+        Mock Resolve-Path { return @{ 'Path' = 'c:/path' } }
+        Get-PodeRelativePath -Path './path' -JoinRoot -Resolve | Should Be 'c:/path'
+    }
+
+    It 'Returns path for a relative path joined to passed root' {
+        Mock Test-PodeRelativePath { return $true }
+        Mock Join-Path { return 'e:/path' }
+        Get-PodeRelativePath -Path './path' -JoinRoot -RootPath 'e:/' | Should Be 'e:/path'
     }
 }