+++ tests/unit/Tools/Middleware.Tests.ps1
@@ -345,6 +345,27 @@ Describe 'Get-PodeRouteValidateMiddleware' {
         }) | Should Be $true
     }
 
+    It 'Returns a ScriptBlock and invokes it as true, overriding the content type' {
+        $WebEvent = @{ 'Parameters' = @{}; 'ContentType' = 'text/plain' }
+
+        $r = Get-PodeRouteValidateMiddleware
+        $r.Name | Should Be '@route-valid'
+        $r.Logic | Should Not Be $null
+
+        Mock Get-PodeRoute { return @{
+            'Parameters' = @{};
+            'Logic' = { Write-Host 'hello' };
+            'ContentType' = 'application/json';
+        } }
+
+        (. $r.Logic @{
+            'Method' = 'GET';
+            'Path' = '/';
+        }) | Should Be $true
+
+        $WebEvent.ContentType | Should Be 'application/json'
+    }
+
     It 'Returns a ScriptBlock and invokes it as false' {
         $r = Get-PodeRouteValidateMiddleware
         $r.Name | Should Be '@route-valid'
@@ -430,10 +451,28 @@ Describe 'Get-PodePublicMiddleware' {
         $r.Name | Should Be '@public'
         $r.Logic | Should Not Be $null
 
-        Mock Get-PodeStaticRoutePath { return $null }
+        Mock Get-PodeStaticRoutePath { return @{ 'Path' = $null } }
         (. $r.Logic @{
             'Path' = '/'; 'Protocol' = 'http'; 'Endpoint' = '';
         }) | Should Be $true
+    }
+
+    It 'Returns a ScriptBlock, invokes false for static path, flagged as download' {
+        $r = Get-PodePublicMiddleware
+        $r.Name | Should Be '@public'
+        $r.Logic | Should Not Be $null
+
+        $PodeContext = @{ 'Server' = @{
+            'Web' = @{ 'Static' = @{ } }
+        }}
+
+        Mock Get-PodeStaticRoutePath { return @{ 'Path' = '/'; 'Download' = $true } }
+        Mock Attach { }
+        (. $r.Logic @{
+            'Path' = '/'; 'Protocol' = 'http'; 'Endpoint' = '';
+        }) | Should Be $false
+
+        Assert-MockCalled Attach -Times 1 -Scope It
     }
 
     It 'Returns a ScriptBlock, invokes false for static path, with no caching' {
@@ -449,11 +488,13 @@ Describe 'Get-PodePublicMiddleware' {
             }}
         }}
 
-        Mock Get-PodeStaticRoutePath { return '/' }
-        Mock Write-PodeValueToResponseFromFile { }
+        Mock Get-PodeStaticRoutePath { return @{ 'Path' = '/' } }
+        Mock File { }
         (. $r.Logic @{
             'Path' = '/'; 'Protocol' = 'http'; 'Endpoint' = '';
         }) | Should Be $false
+
+        Assert-MockCalled File -Times 1 -Scope It
     }
 
     It 'Returns a ScriptBlock, invokes false for static path, with no caching from exclude' {
@@ -470,11 +511,13 @@ Describe 'Get-PodePublicMiddleware' {
             }}
         }}
 
-        Mock Get-PodeStaticRoutePath { return '/' }
-        Mock Write-PodeValueToResponseFromFile { }
+        Mock Get-PodeStaticRoutePath { return @{ 'Path' = '/' } }
+        Mock File { }
         (. $r.Logic @{
             'Path' = '/'; 'Protocol' = 'http'; 'Endpoint' = '';
         }) | Should Be $false
+
+        Assert-MockCalled File -Times 1 -Scope It
     }
 
     It 'Returns a ScriptBlock, invokes false for static path, with no caching from include' {
@@ -491,11 +534,13 @@ Describe 'Get-PodePublicMiddleware' {
             }}
         }}
 
-        Mock Get-PodeStaticRoutePath { return '/' }
-        Mock Write-PodeValueToResponseFromFile { }
+        Mock Get-PodeStaticRoutePath { return @{ 'Path' = '/' } }
+        Mock File { }
         (. $r.Logic @{
             'Path' = '/'; 'Protocol' = 'http'; 'Endpoint' = '';
         }) | Should Be $false
+
+        Assert-MockCalled File -Times 1 -Scope It
     }
 
     It 'Returns a ScriptBlock, invokes false for static path, with caching' {
@@ -511,10 +556,12 @@ Describe 'Get-PodePublicMiddleware' {
             }}
         }}
 
-        Mock Get-PodeStaticRoutePath { return '/' }
-        Mock Write-PodeValueToResponseFromFile { }
+        Mock Get-PodeStaticRoutePath { return @{ 'Path' = '/' } }
+        Mock File { }
         (. $r.Logic @{
             'Path' = '/'; 'Protocol' = 'http'; 'Endpoint' = '';
         }) | Should Be $false
+
+        Assert-MockCalled File -Times 1 -Scope It
     }
 }