+++ docs/release-notes.md
@@ -1,5 +1,22 @@
 # Release Notes
 
+## v0.29.0
+
+```plain
+### Enhancements
+* #216: Multi-content-type support on Error Pages
+* #232: Support for setting/forcing default content types on routes
+* #243: Support on Static Routes to flag them as "Download Only"
+* #248: Ability to alter the server's root path
+
+### Packaging
+* #227: Docker images updated to PowerShell Core v6.2.0
+* #233: New ARM32 docker image, enabling support for Raspberry Pi
+
+### Build
+* #237, #238, #239: Updates to CI tools
+```
+
 ## v0.28.1
 
 ```plain